Add files via upload
his data set contains booking information for hotels. It includes information such as booking dates, type of hotels, date of reservation, when the booking was made, length of stay, and customer information. To run the models and reduce the complexity of the categorical target variable, I reduced the rows to 3999, which provide booking information from July 2015 to February 2016. The target variable I have chosen is categorical. The business question is: which type of reservation platform is most preferred for booking? The target variable is Type_of_Reservation. With this analysis, the management could decide on postings of room availability on online platforms, manage direct, travel agent, and corporate bookings, vary prices, and manage their bookings to earn profits. The business value associated with it is that people around the world stay at hotels during vacations, business trips, get-togethers, and so on. The increase in technology has introduced various online platforms to make hotel reservations. Hotels can get reservations from various platforms, the reservations could be made through a travel agent, a travel organization, a corporate organization, or in person. If the management of a hotel knows which is the most preferred platform for making a reservation, they could divide their reservation openings between the reservation sectors and reserve rooms.